+++ app/decorators/document_decorator.rb
@@ -11,17 +11,21 @@ class DocumentDecorator < Draper::Decorator
   #   end
 
   def file_link
-  	h.link_to model.title , model.attached_doc.url
+  	h.link_to model.title , model.attached_doc.url if model.attached_doc && h.can?(:read , model)
   end
 
   def edit_link
-    h.tag_label h.link_to("Edit" , h.edit_lesson_document_path(object.lesson,object)) , :success
+    if h.can? :update , model.lesson
+      h.tag_label h.link_to("Edit" , h.edit_lesson_document_path(object.lesson,object)) , :success
+    end
   end
 
   def destroy_link remote = false
-    h.tag_label h.link_to("Delete" , h.lesson_document_path(object.lesson,object) , 
+    if h.can? :update , model.lesson
+      h.tag_label h.link_to("Delete" , h.lesson_document_path(object.lesson,object) , 
                           method: :delete , remote: remote,
                           data: {confirm: "Are you sure?"}) , :danger
+    end
   end
 
 end
